requirements.txt
asgiref==3.4.1
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
